+++ pypeit/reduce.py
@@ -559,6 +559,12 @@ class Reduce(object):
             msgs.info("Global sky subtraction for slit: {:d}".format(slit_idx))
             thismask = self.slitmask == slit_spat
             inmask = (self.sciImg.fullmask == 0) & thismask & skymask_now
+            # All masked?
+            if not np.any(inmask):
+                msgs.warn("No pixels for fitting sky.  If you are using mask_by_boxcar=True, your radius may be too large.")
+                self.reduce_bpm[slit_idx] = True
+                continue
+
             # Find sky
             self.global_sky[thismask] = skysub.global_skysub(self.sciImg.image, self.sciImg.ivar, self.tilts,
                                                              thismask, self.slits_left[:,slit_idx],
@@ -971,6 +977,12 @@ class MultiSlitReduce(Reduce):
         # Instantiate the specobjs container
         sobjs = specobjs.SpecObjs()
 
+        # Masking options
+        if self.par['reduce']['skysub']['mask_by_boxcar']:
+            boxcar_rad_skymask = self.par['reduce']['extraction']['boxcar_radius'] / self.get_platescale(None),
+        else:
+            boxcar_rad_skymask = None
+
         # Loop on slits
         for slit_idx in gdslits:
             slit_spat = self.slits.spat_id[slit_idx]
@@ -1002,6 +1014,7 @@ class MultiSlitReduce(Reduce):
                                 cont_fit=self.par['reduce']['findobj']['find_cont_fit'],
                                 npoly_cont=self.par['reduce']['findobj']['find_npoly_cont'],
                                 fwhm=self.par['reduce']['findobj']['find_fwhm'],
+                                boxcar_rad_skymask=boxcar_rad_skymask,
                                 maxdev=self.par['reduce']['findobj']['find_maxdev'],
                                 find_min_max=self.par['reduce']['findobj']['find_min_max'],
                                 qa_title=qa_title, nperslit=self.par['reduce']['findobj']['maxnumber'],
@@ -1084,7 +1097,8 @@ class MultiSlitReduce(Reduce):
                     bsp=self.par['reduce']['skysub']['bspline_spacing'],
                     sn_gauss=self.par['reduce']['extraction']['sn_gauss'],
                     show_profile=show_profile,
-                    use_2dmodel_mask=self.par['reduce']['extraction']['use_2dmodel_mask'])
+                    use_2dmodel_mask=self.par['reduce']['extraction']['use_2dmodel_mask'],
+                    no_local_sky=self.par['reduce']['skysub']['no_local_sky'])
 
         # Set the bit for pixels which were masked by the extraction.
         # For extractmask, True = Good, False = Bad
@@ -1209,6 +1223,8 @@ class EchelleReduce(Reduce):
             max_snr=self.par['reduce']['findobj']['ech_find_max_snr'],
             min_snr=self.par['reduce']['findobj']['ech_find_min_snr'],
             nabove_min_snr=self.par['reduce']['findobj']['ech_find_nabove_min_snr'],
+            skymask_by_boxcar=self.par['reduce']['skysub']['mask_by_boxcar'],
+            boxcar_rad=self.par['reduce']['extraction']['boxcar_radius'],  # arcsec
             show_trace=show_trace, debug=debug)
 
         # Steps